Work in progress on lazy evaluation of query results.

+++ structr-neo4j-bolt-driver/src/main/java/org/structr/bolt/index/AbstractCypherIndex.java
@@ -95,7 +95,7 @@ public abstract class AbstractCypherIndex<T extends PropertyContainer> implement
 		CONVERTERS.put(Double.class,  new DoubleTypeConverter());
 	}
 
-	protected final FixedSizeCache<Integer, Long> resultCountCache;
+	protected final FixedSizeCache<Integer, Integer> resultCountCache;
 	protected final BoltDatabaseService db;
 
 	public AbstractCypherIndex(final BoltDatabaseService db, final int queryCacheSize) {
